+++ src/molecules/GasReceivedChart/GasReceivedChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Skeleton from '../../atoms/Skeleton/Skeleton';
+import SkeletonChart from '../../molecules/SkeletonChart/SkeletonChart';
 import Message from '../../atoms/Message/Message';
 
 import { useQuery, gql } from '@apollo/client';
@@ -23,9 +23,8 @@ const GAS_EXCHANGES_CHART = gql`
 	}
 `;
 
-const now = DateTime.now();
-
 export default function GasReceivedChart() {
+	const now = DateTime.now();
 	const { loading, error, data } = useQuery(GAS_EXCHANGES_CHART, {
 		variables: {
 			resolution: 'TWO_HOURS',
@@ -36,7 +35,7 @@ export default function GasReceivedChart() {
 		}
 	});
 
-	if (loading) return <Skeleton width="100%" height="300px" />;
+	if (loading) return <SkeletonChart />;
 	if (error) return <Message type="error">{error.message}</Message>;
 
 	return (
@@ -56,7 +55,7 @@ export default function GasReceivedChart() {
 						lineColor: 'hsla(var(--color-secondary-shade-3-h), var(--color-secondary-shade-3-s), var(--color-secondary-shade-3-l), .4)',
 						tickColor: 'hsla(var(--color-secondary-shade-3-h), var(--color-secondary-shade-3-s), var(--color-secondary-shade-3-l), .4)',
 						plotBands: data.gasExchanges.filter(exchange => exchange.dataPointsCount === 0).map(exchange => ({
-							color: 'hsla(var(--color-secondary-shade-1-h), var(--color-secondary-shade-1-s), var(--color-secondary-shade-1-l), 0.1)', // Color value
+							color: 'hsla(var(--color-secondary-shade-1-h), var(--color-secondary-shade-1-s), var(--color-secondary-shade-1-l), 0.2)', // Color value
 							from: exchange.period.start * 1000,
 							to: exchange.period.end * 1000
 						})),