+++ src/molecules/WaterReceivedWeekChart/fragments.js
@@ -1,0 +1,11 @@
+import { gql } from '@apollo/client';
+
+export const CONSUMPTION = gql`
+	fragment consumption on WaterConsumption {
+		received
+		period {
+			start
+			end
+		}
+	}
+`;